+++ src/extension.ts
@@ -1,6 +1,9 @@
 import * as vscode from 'vscode'
+import { FindingsOverviewProvider } from './providers/findingsOverviewProvider'
+import { FindingsTreeProvider } from './providers/findingsTreeProvider'
+import { SettingsTreeProvider } from './providers/settingsTreeProvider'
+import { exportToJSON, showHTMLReport } from './report'
 import { Settings } from './settings'
-import { showSettingsPanel } from './settingsPanel'
 import { Watchtower } from './watchtower'
 
 
@@ -9,29 +12,40 @@ export function activate(context: vscode.ExtensionContext) {
 	console.log("Watchtower Extension Loading")
 
 	const settings = Settings.getInstance(context)
-	const watchtower = Watchtower.getInstance(context.extensionUri)
 
+	const findingsTree = new FindingsTreeProvider()
+	const findingsOverview = new FindingsOverviewProvider()
+	const settingsTree = new SettingsTreeProvider()
+
+	const watchtower = Watchtower.getInstance(findingsTree, findingsOverview, settingsTree)
+
+
+	/////////////////////////////
+	// Views
+	/////////////////////////////
+
+	context.subscriptions.push(
+		vscode.window.registerWebviewViewProvider(FindingsTreeProvider.viewType, findingsTree)
+	)
+	context.subscriptions.push(
+		vscode.window.registerWebviewViewProvider(FindingsOverviewProvider.viewType, findingsOverview)
+	)
+	context.subscriptions.push(
+		vscode.window.registerTreeDataProvider(SettingsTreeProvider.viewType, settingsTree)
+	)
 	/////////////////////////////
 	// Commands
 	/////////////////////////////
 	const commands = [
 		vscode.commands.registerCommand('watchtower.scan', watchtower.commandRunScan.bind(watchtower)),
-		vscode.commands.registerCommand('watchtower.disableWorkspaceStartupScan', watchtower.commandDisableStartupScanForWorkspace.bind(watchtower)),
-		vscode.commands.registerCommand('watchtower.disableWorkspaceRealTimeDetection', watchtower.commandDisableRealTimeDetectionForWorkspace.bind(watchtower)),
-		vscode.commands.registerCommand('watchtower.enableWorkspaceStartupScan', () => settings.setWorkspaceStartupScan(true)),
-		vscode.commands.registerCommand('watchtower.enableWorkspaceRealTimeDetection', () => settings.setWorkspaceRealTimeDetection(true)),
-		vscode.commands.registerCommand('watchtower.showSettings', showSettingsPanel),
+		vscode.commands.registerCommand('watchtower.toggleWorkspaceStartupScan', watchtower.commandToggleWorkspaceStartupScan.bind(watchtower)),
+		vscode.commands.registerCommand('watchtower.toggleWorkspaceRealTime', watchtower.commandToggleWorkspaceRealTime.bind(watchtower)),
+		vscode.commands.registerCommand('watchtower.exportToJSON', () => exportToJSON(watchtower.findings, false)),
+		vscode.commands.registerCommand('watchtower.showReport', () => showHTMLReport(watchtower.findings, context.extensionUri)),
+		vscode.commands.registerCommand('watchtower.revealFinding', (finding) => findingsTree.revealFinding(finding)),
 	]
 
 	commands.forEach(command => context.subscriptions.push(command))
-
-
-
-	// If startup scans are off globally and no project overrides, skip all automatic scanning
-	if (settings.shouldEnforceRestrictedScanOnlySetting()) {
-		console.log('Watchtower: startupScans is set to OnUntrusted and workspace is trusted — skipping scans')
-		return
-	}
 
 
 	/////////////////////////////
@@ -48,13 +62,15 @@ export function activate(context: vscode.ExtensionContext) {
 		watcher
 	]
 
-	realTimeListeners.forEach(listener => context.subscriptions.push(listener))
+	if (settings.shouldRunRealtimeScanForWorkspace()) {
+		realTimeListeners.forEach(listener => context.subscriptions.push(listener))
+	}
 
 
 	context.subscriptions.push(vscode.workspace.onDidGrantWorkspaceTrust(() => {
-		// if user only wants scans on untrusted workspaces, and they just granted trust, dispose real-time listeners
 		console.log("Workspace trust granted")
-		if (settings.shouldEnforceRestrictedScanOnlySetting()) {
+		settingsTree.refresh()
+		if (!settings.shouldRunRealtimeScanForWorkspace()) {
 			console.log("Disposing real time listeners due to workspace trust change")
 			realTimeListeners.forEach(listener => listener.dispose())
 		}
@@ -63,6 +79,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 	watchtower.runInitialScan()
+
 
 	if (vscode.window.activeTextEditor)
 		watchtower.onActiveEditorChanged(vscode.window.activeTextEditor)